### 2017-12-22 #519 @whedon
[PRE REVIEW]: reper - Genome-wide identification, classification and quantification of repetitive elements without an assembled genome
**Submitting author:** @nterhoeven (<a href="http://orcid.org/0000-0002-8669-0000">Niklas Terhoeven</a>)
**Repository:** <a href="https://github.com/nterhoeven/reper" target ="_blank">https://github.com/nterhoeven/reper</a>
**Version:** v1.0.0
