### 2017-12-22 a24b97f
Fixing bibtex name
