+++ paper/paper.bib
@@ -222,7 +222,7 @@
 	file = {evv050.pdf:/home/niklas/Zotero/storage/Z2PUPMEQ/evv050.pdf:application/pdf}
 }
 
-@article{camacho_blast+:_2009,
+@article{camacho_blast,
 	title = {{BLAST}+: architecture and applications},
 	volume = {10},
 	issn = {1471-2105},
