### 2017-12-22 #519 @arfon
@nterhoeven - could you please merge this PR which fixes up some issues with your paper: https://github.com/nterhoeven/reper/pull/4
