### 2018-01-07 #519 @jsgalan
Hi @nterhoeven and all, i reviewed the article and have minor comments.

However, when i wanted to install the software through Docker, had some problems. I am working in osx and the installer it seems is made for ubuntu, attached is the error log for the build. Any of you guys know how to get around this?

Let me know if I am doing something wrong.

Best

[reper docker build.txt](https://github.com/openjournals/joss-reviews/files/1610196/reper.docker.build.txt)
