### 2018-01-08 #519 @nterhoeven
Hi @jsgalan thanks for reviewing. Would you mind to [open an issue](https://github.com/nterhoeven/reper/issues/new) for your comments. I will then take a look at this.

The Docker error is strange - Docker should be platform independent. Looking at the log, I see the following error:

```
Err:1 http://archive.ubuntu.com/ubuntu xenial InRelease
  Temporary failure resolving archive.ubuntu.com'
```
This seems to be a network issue on your system. maybe related to the DNS server you are using. take a look at [this stackexchange question](https://serverfault.com/questions/386918/apt-get-cannot-connect-failure-resolving-archive-ubuntu-com#386921)

Can you please try to use the automatically built container from dockerhub? This should also be faster than building the container on a local machine.

```
docker pull nterhoeven/reper
```
