### 2018-01-08 #519 @mgymrek
Hi @jsgalan thanks for agreeing to review! Sorry for not seeing this earlier. I will now start a review thread where you can paste additional comments and should go through the reviewer checklist.
