### 2018-01-08 #519 @mgymrek
@jsgalan please see review thread linked above where we can continue discussion. I am closing this issue for now.
