### 2018-01-11 #527 @jsgalan
I still have major problems with the docker installation.

I am blaming mostly the osx High Sierra docker for Mac for the inconvenience.

Any other reviewer using osx High Sierra and Docker could help?

Best
