### 2018-01-11 #527 @nterhoeven
Hi @jsgalan that's a pity.

As it seems, there are other poeple who have problems with Docker after upgrading to High Sierra. For some [reinstalling docker seems to work](https://github.com/docker/for-mac/issues/2091). I am not a Mac expert, but maybe it is worth a try.

If docker is not working for you, you can take a look at [singularity](http://singularity.lbl.gov). This is a similar tool. Take a look at the [reper wiki](https://github.com/nterhoeven/reper/wiki/Installation#singularity) for instructions to run reper with singularity.

Message me if you have any questions.
