### 2018-01-11 10aad63
Reformated apt install line
For better reading and easier handling of package installation, I basically wrap the line into a multi line command
