### 2018-01-11 948f7bd
Reformated apt install line
For better reading and easier handling of package installation, I basically wrap the line into a multi line command
