### 2018-01-11 081404f
Wraps all installation lines into multi line RUN commands
