+++ Dockerfile
@@ -30,26 +30,70 @@ COPY scripts/ ./scripts/
 ENV PATH="$reperDir":$PATH
 
 WORKDIR $depDir
-RUN wget https://github.com/gmarcais/Jellyfish/releases/download/v2.2.6/jellyfish-2.2.6.tar.gz && tar xzf jellyfish-2.2.6.tar.gz && cd jellyfish-2.2.6 && ./configure && make && make install
-RUN wget https://github.com/trinityrnaseq/trinityrnaseq/archive/Trinity-v2.4.0.tar.gz && tar xzf Trinity-v2.4.0.tar.gz && cd trinityrnaseq-Trinity-v2.4.0 && make && make plugins
+RUN wget https://github.com/gmarcais/Jellyfish/releases/download/v2.2.6/jellyfish-2.2.6.tar.gz && \
+    tar xzf jellyfish-2.2.6.tar.gz && \
+    cd jellyfish-2.2.6 && \
+    ./configure && \
+    make && \
+    make install
+
+RUN wget https://github.com/trinityrnaseq/trinityrnaseq/archive/Trinity-v2.4.0.tar.gz && \
+    tar xzf Trinity-v2.4.0.tar.gz && \
+    cd trinityrnaseq-Trinity-v2.4.0 && \
+    make && \
+    make plugins
 ENV TRINITY_HOME="$depDir"/trinityrnaseq-Trinity-v2.4.0
-RUN wget https://github.com/weizhongli/cdhit/releases/download/V4.6.7/cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz && tar xzf cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz
-RUN wget https://github.com/BenLangmead/bowtie2/releases/download/v2.3.2/bowtie2-2.3.2-linux-x86_64.zip && unzip bowtie2-2.3.2-linux-x86_64.zip
-RUN wget https://github.com/samtools/samtools/releases/download/1.4.1/samtools-1.4.1.tar.bz2 && tar xjf samtools-1.4.1.tar.bz2 && cd samtools-1.4.1 && ./configure && make && make install
-RUN wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.2.28/ncbi-blast-2.2.28+-x64-linux.tar.gz && tar xzf ncbi-blast-2.2.28+-x64-linux.tar.gz
+
+RUN wget https://github.com/weizhongli/cdhit/releases/download/V4.6.7/cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz && \
+    tar xzf cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz
+
+RUN wget https://github.com/BenLangmead/bowtie2/releases/download/v2.3.2/bowtie2-2.3.2-linux-x86_64.zip && \
+    unzip bowtie2-2.3.2-linux-x86_64.zip
+
+RUN wget https://github.com/samtools/samtools/releases/download/1.4.1/samtools-1.4.1.tar.bz2 && \
+    tar xjf samtools-1.4.1.tar.bz2 && \
+    cd samtools-1.4.1 && \
+    ./configure && \
+    make && \
+    make install
+
+RUN wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.2.28/ncbi-blast-2.2.28+-x64-linux.tar.gz && \
+    tar xzf ncbi-blast-2.2.28+-x64-linux.tar.gz
+
 RUN git clone https://github.com/thackl/kmer-scripts.git
 
 WORKDIR $depDir/lib
 ENV PERL5LIB="$depDir/lib:$PERL5LIB"
[…]
