### 2018-01-11 bb02134
Removed unnecessary group from chmod
