### 2018-01-11 1a81310
Remove dist-upgrade to shrink image size
