+++ Dockerfile
@@ -1,7 +1,6 @@
 FROM ubuntu:16.04
 
 RUN apt-get update && \
-    apt-get -y dist-upgrade && \
     apt-get -y install \
        wget \
        g++ \
