+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM ubuntu:16.04
 
 RUN apt-get update && \
-    apt-get -y install \
+    apt-get install --yes \
        wget \
        g++ \
        build-essential \
