### 2018-01-11 3de3505
Skip installation of recommended packages to further shrink the image
