+++ Dockerfile
@@ -1,7 +1,7 @@
 FROM ubuntu:16.04
 
 RUN apt-get update && \
-    apt-get install --yes \
+    apt-get install --yes --no-install-recommends \
        wget \
        g++ \
        build-essential \
