### 2018-01-11 e789def
Remove downloaded content from apt update to shrink the image
