### 2018-01-11 2e750d2
Remove emacs package to shrink the image
