### 2018-01-11 b59b893
Removing temporary files after jellyfish installation
