### 2018-01-11 9ca0798
Fix shared library issue for jellyfish
Solves #5
