### 2018-01-11 4f0e71a
Calling tests for Log4perl installation
