### 2018-01-11 9a381c1
Removing temporary files after Log4perl installation
