+++ Dockerfile
@@ -77,6 +77,13 @@ RUN wget http://search.cpan.org/CPAN/authors/id/M/MS/MSCHILLI/Log-Log4perl-1.49.
     cd .. && \
     rm -rf Log-Log4perl-1.49 Log-Log4perl-1.49.tar.gz
 
+RUN wget http://search.cpan.org/CPAN/authors/id/P/PL/PLICEASE/File-Which-1.22.tar.gz && \
+    tar xzf File-Which-1.22.tar.gz && \
+    cd File-Which-1.22 && \
+    perl Makefile.PL && \
+    make && \
+    make install
+
 RUN git clone https://github.com/BioInf-Wuerzburg/perl5lib-Fastq.git && \
     mv perl5lib-Fastq/lib/* .
 
@@ -85,13 +92,6 @@ RUN git clone https://github.com/BioInf-Wuerzburg/perl5lib-Fasta.git && \
 
 RUN git clone https://github.com/thackl/perl5lib-Jellyfish.git && \
     mv perl5lib-Jellyfish/lib/* .
-
-RUN wget http://search.cpan.org/CPAN/authors/id/P/PL/PLICEASE/File-Which-1.22.tar.gz && \
-    tar xzf File-Which-1.22.tar.gz && \
-    cd File-Which-1.22 && \
-    perl Makefile.PL && \
-    make && \
-    make install
 
 RUN git clone https://github.com/thackl/perl5lib-Kmer.git && \
     mv perl5lib-Kmer/lib/* .
