### 2018-01-11 0a37f95
Removing temporary files after File::Which installation
