### 2018-01-11 3eb36bb
Install other required perl modules in a loop and delete the github clone for shrinking
