### 2018-01-11 148eb44
Calling tests for samtools installation
