+++ Dockerfile
@@ -57,6 +57,7 @@ RUN wget https://github.com/samtools/samtools/releases/download/1.4.1/samtools-1
     cd samtools-1.4.1 && \
     ./configure && \
     make && \
+    make check && \
     make install
 
 RUN wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.2.28/ncbi-blast-2.2.28+-x64-linux.tar.gz && \
