### 2018-01-11 a711539
Removing temporary files after samtools installation
