### 2018-01-11 91292ca
Reorder bowtie2 installation
