### 2018-01-11 3698065
Removing temporary files after bowtie2 installation
