### 2018-01-11 db731a9
Include bowtie2 installation to PATH
