### 2018-01-11 3aac54f
Installing cd-hit and removing temporary files after cd-hit installation
