+++ Dockerfile
@@ -52,7 +52,11 @@ RUN wget https://github.com/trinityrnaseq/trinityrnaseq/archive/Trinity-v2.4.0.t
 ENV TRINITY_HOME="$depDir"/trinityrnaseq-Trinity-v2.4.0
 
 RUN wget https://github.com/weizhongli/cdhit/releases/download/V4.6.7/cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz && \
-    tar xzf cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz
+    tar xzf cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz && \
+    cd cd-hit-v4.6.7-2017-0501 && \
+    make install && \
+    cd .. && \
+    rm -rf cd-hit-v4.6.7-2017-0501 cd-hit-v4.6.7-2017-0501-Linux-binary.tar.gz
 
 RUN wget https://github.com/samtools/samtools/releases/download/1.4.1/samtools-1.4.1.tar.bz2 && \
     tar xjf samtools-1.4.1.tar.bz2 && \
