### 2018-01-11 2616451
Reorder trinity installation
