### 2018-01-11 2851318
Removing temporary files after blast+ installation
