+++ Dockerfile
@@ -62,7 +62,8 @@ RUN wget https://github.com/samtools/samtools/releases/download/1.4.1/samtools-1
     rm -rf samtools-1.4.1 samtools-1.4.1.tar.bz2
 
 RUN wget ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/2.2.28/ncbi-blast-2.2.28+-x64-linux.tar.gz && \
-    tar xzf ncbi-blast-2.2.28+-x64-linux.tar.gz
+    tar xzf ncbi-blast-2.2.28+-x64-linux.tar.gz && \
+    rm ncbi-blast-2.2.28+-x64-linux.tar.gz
 
 RUN wget https://github.com/trinityrnaseq/trinityrnaseq/archive/Trinity-v2.4.0.tar.gz && \
     tar xzf Trinity-v2.4.0.tar.gz && \
