### 2018-01-11 de14a7d
Include blast+ installation to PATH
