### 2018-01-11 01633a4
Removing temporary files after trinity installation
