### 2018-01-11 3203e70
Changing mod for $depDir for each layer to shrink image
