### 2018-01-11 43e90c4
Skip mod for $depDir for each layer to shrink image
