### 2018-01-12 5f16e47
Add update_exe_path.sh script to update reper.conf file
