### 2018-01-12 7f564d9
Add trinity to PATH variable
