### 2018-01-12 5c9c2d9
Add kmer-scripts to PATH
