### 2018-01-12 752b20c
Fix update_exe_path.sh to work with a parameter instead of fixed location
