### 2018-01-12 578fd2d
Switched to headless jre instead of graphics supporting jre
