### 2018-01-12 98aef67
fix docker status badge commands
