+++ README.md
@@ -1,8 +1,8 @@
 # reper - Genome-wide identification, classification and quantification of repetitive elements without an assembled genome
 
 
-[![Docker Automated build](https://img.shields.io/docker/automated/jrottenberg/ffmpeg.svg)](https://hub.docker.com/r/nterhoeven/reper/)
-[![Docker Build Status](https://img.shields.io/docker/build/jrottenberg/ffmpeg.svg)](https://hub.docker.com/r/nterhoeven/reper/)
+[![Docker Automated build](https://img.shields.io/docker/automated/nterhoeven/reper.svg)](https://hub.docker.com/r/nterhoeven/reper/)
+[![Docker Build Status](https://img.shields.io/docker/build/nterhoeven/reper.svg)](https://hub.docker.com/r/nterhoeven/reper/)
 [![DOI](https://zenodo.org/badge/80427752.svg)](https://zenodo.org/badge/latestdoi/80427752)
 [![status](http://joss.theoj.org/papers/f0d16a43d8b031695f151ea25e0d47b0/status.svg)](http://joss.theoj.org/papers/f0d16a43d8b031695f151ea25e0d47b0)
 
