### 2018-01-12 nterhoeven/reper#9 @nterhoeven
installation problems on Mac
@jsgalan reported an issue with running reper inside docker on a Mac running High Sierra (https://github.com/openjournals/joss-reviews/issues/519)
