### 2018-01-12 #527 @jsgalan
Hi @nterhoeven i have been trying with Singularity, i got a few than with Docker, still somehow the VM crashed and wasnt able to restart it, I will try it today again.
