### 2018-01-12 nterhoeven/reper#9 @nterhoeven
We updated the dockerfile for reper. Currently a new build is running on dockerhub.

@jsgalan: Would you mind running a `docker pull nterhoeven/reper` when it's finished and report back if it works. We found a colleague running High Sierra and it worked on his machine.
