### 2018-01-16 nterhoeven/reper#9 @greatfireball
One of my colleagues was able to build the image and run a container with the current version of the Dockerfile. Originally we had some memory issues running another Docker container [here](https://github.com/chloroExtractorTeam/chloroExtractor/issues/81#issuecomment-349329194), but he solved the memory issue and it ran (his configuration is stated in his comment). AFAIK that was all he changed on his configuration.
