### 2018-01-16 nterhoeven/reper#9 @DanielAmsel
Hi *
Since I was offered a new Docker Version today,
I re-ran everything and it still works for me (see commands below or in wiki).
My setup/configuration/system:

Docker Version 17.12.0-ce-mac47
Channel: stable
CPUs: 2
Memory: 4.0 GB

macOS High Sierra
Version 10.13.2
MacBook Pro Early 2015
2.7GHz Intel Core i5
8GB 1867 MHz DDR3

Commands:
`docker pull nterhoeven/reper`
`docker run --rm -it nterhoeven/reper`

Alternatively:
`git clone https://github.com/nterhoeven/reper.git`
`cd reper`
`docker build .`
Finished with message of successful build.

Kind regards,
Daniel
