### 2018-02-05 #527 @jsgalan
Hi @nterhoeven

The paper and implementation are excellent. A few corrections are needed:

- (kmer counts based on jelly sh (Marçais and Kingsford 2011)) <-- check parenthesis use here
-To create exemplar sequences of each repeat in the genome, the assembled repeats ar clustered using cd-hit ((W. Li and Godzik 2006),(Fu et al. 2012)). <-- should read "repeats are clustered"

Hope this finishes the review process

Best
