### 2018-02-05 #527 @mgymrek
Thanks @jsgalan!
@nterhoeven after you make these minor changes update the thread and then we should be ready to go.
