### 2018-02-06 6c9e2f1
fix parentheses in citations
