### 2018-02-06 #527 @nterhoeven
Hi @jsgalan and @mgymrek,

thanks for the comments :-) I created an updated version of the paper.md file.

How are we going to proceed? I believe I have to create a new release and provide you with a DOI for it. Is this correct?
