### 2018-02-07 #527 @nterhoeven
Hi @arfon,
I created a new release (v1.1.0). Here is the new DOI: 10.5281/zenodo.1167969
